Move app creation and jwt initialization
Move app creation into its factory functiona and move jwt initialization
into a module for readability
