+++ backend/__init__.py
@@ -1,5 +1,13 @@
-from flask import Flask
+##app = Flask(__name__)
 
-app = Flask(__name__)
+from backend.app import create_app
 
-import backend.views
+app = create_app()
+
+
+import backend.jwt
+#from backend.jwt import create_jwt
+
+#create_jwt(app)
+
+#import backend.jwt
